+++ src/main/java/de/samply/reporter/report/ReportGenerator.java
@@ -2,6 +2,7 @@ package de.samply.reporter.report;
 
 import de.samply.reporter.app.ReporterConst;
 import de.samply.reporter.context.CellContext;
+import de.samply.reporter.context.CellStyleContext;
 import de.samply.reporter.context.Context;
 import de.samply.reporter.context.ContextGenerator;
 import de.samply.reporter.exporter.ExporterClient;
@@ -16,6 +17,8 @@ import de.samply.reporter.template.SheetTemplate;
 import de.samply.reporter.template.script.Script;
 import de.samply.reporter.utils.ExternalSheetUtils;
 import de.samply.reporter.utils.ExternalSheetUtilsException;
+import de.samply.reporter.utils.FileUtils;
+import de.samply.reporter.utils.PercentageLogger;
 import de.samply.reporter.zip.ExporterUnzipper;
 import de.samply.reporter.zip.ExporterUnzipperException;
 import java.io.FileOutputStream;
@@ -35,6 +38,8 @@ import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.util.CellRangeAddress;
 import org.apache.poi.xssf.streaming.SXSSFSheet;
 import org.apache.poi.xssf.streaming.SXSSFWorkbook;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.util.FileSystemUtils;
@@ -42,6 +47,7 @@ import org.springframework.util.FileSystemUtils;
 @Component
 public class ReportGenerator {
 
+  private final static Logger logger = LoggerFactory.getLogger(ReportGenerator.class);
   private final ExporterClient exporterClient;
   private final ExporterUnzipper exporterUnzipper;
   private final ScriptEngineManager scriptEngineManager;
@@ -74,15 +80,24 @@ public class ReportGenerator {
   }
 
   private void generate(ReportTemplate template, String filePath, ReportMetaInfo reportMetaInfo) {
+    logger.info("Extracting paths...");
     Path[] paths = extractPaths(filePath);
+    logger.info("Generating context...");
     Context context = contextGenerator.generate(template, paths);
-    Map<Script, ScriptResult> scriptResultMap = scriptEngineManager.generateRawQualityReport(
+    logger.info("Generating raw report...");
+    Map<Script, ScriptResult> scriptResultMap = scriptEngineManager.generateRawReport(
         template, context);
+    logger.info("Generating excel file");
     Workbook workbook = new SXSSFWorkbook(workbookWindow);
+    logger.info("Filling excel file with data...");
     fillWorkbookWithData(workbook, template, scriptResultMap);
-    addFormatToWorkbook(workbook, template);
+    logger.info("Adding format to excel file...");
+    addFormatToWorkbook(workbook, template, context);
+    logger.info("Writing excel file...");
     writeWorkbook(reportMetaInfo.path(), workbook);
+    logger.info("Removing temporal files...");
     removeTemporalFiles(paths[0].getParent(), scriptResultMap.values());
+    logger.info("Excel file generated satisfactory.");
   }
 
   private void removeTemporalFiles(Path sourceFilesDirectory,
@@ -100,7 +115,7 @@ public class ReportGenerator {
     FileSystemUtils.deleteRecursively(sourceFilesDirectory);
     scriptResults.forEach(scriptResult -> {
       try {
-        Files.delete(scriptResult.getRawResult());
+        Files.delete(scriptResult.rawResult());
       } catch (IOException e) {
         throw new RuntimeException(e);
       }
@@ -127,6 +142,7 @@ public class ReportGenerator {
       Map<Script, ScriptResult> scriptResultMap) {
     template.getSheetTemplates()
         .forEach(sheetTemplate -> {
+          logger.info("Adding sheet '" + sheetTemplate.getName() + "'...");
           if (sheetTemplate.getFileUrl() != null || sheetTemplate.getFilePath() != null) {
             addSheetFromSourceExcelFile(workbook, sheetTemplate);
           } else {
@@ -150,8 +166,10 @@ public class ReportGenerator {
     if (template.getValuesScript() != null) {
       ScriptResult result = scriptResultMap.get(template.getValuesScript().getScript());
       if (result != null) {
-        fillSheetWithData(sheet, template, result);
+        fillSheetWithData(sheet, result);
+        logger.info("Adding autosize to sheet '" + template.getName() + "'...");
         autoSizeSheet(sheet);
+        logger.info("Adding auto filter to sheet '" + template.getName() + "'...");
         addAutoFilter(sheet);
       }
     }
@@ -166,13 +184,16 @@ public class ReportGenerator {
   }
 
   private void createHeaderRow(Workbook workbook, Sheet sheet, SheetTemplate template) {
-    Row row = sheet.createRow(0);
-    AtomicInteger counter = new AtomicInteger(0);
-    template.getColumnTemplates().forEach(
-        columnTemplate -> row.createCell(counter.getAndIncrement())
-            .setCellValue(columnTemplate.getName()));
-    sheet.createFreezePane(0, 1);
-    boldHeaderRow(workbook, row);
+    if (!template.getColumnTemplates().isEmpty()) {
+      logger.info("Creating header row for sheet '" + template.getName() + "'...");
+      Row row = sheet.createRow(0);
+      AtomicInteger counter = new AtomicInteger(0);
+      template.getColumnTemplates().forEach(
+          columnTemplate -> row.createCell(counter.getAndIncrement())
+              .setCellValue(columnTemplate.getName()));
+      sheet.createFreezePane(0, 1);
+      boldHeaderRow(workbook, row);
+    }
   }
 
   private void boldHeaderRow(Workbook workbook, Row titleRow) {
@@ -186,29 +207,33 @@ public class ReportGenerator {
     }
   }
 
-  private void fillSheetWithData(Sheet sheet, SheetTemplate template, ScriptResult result) {
+  private void fillSheetWithData(Sheet sheet, ScriptResult result) {
     try {
-      fillSheetWithDataWithoutExceptionHandling(sheet, template, result);
+      fillSheetWithDataWithoutExceptionHandling(sheet, result);
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
   }
 
-  private void fillSheetWithDataWithoutExceptionHandling(Sheet sheet, SheetTemplate template,
-      ScriptResult result)
+  private void fillSheetWithDataWithoutExceptionHandling(Sheet sheet, ScriptResult result)
       throws IOException {
     AtomicInteger rowIndex = new AtomicInteger(sheet.getLastRowNum() + 1);
-    Files.readAllLines(result.getRawResult())
+    PercentageLogger percentageLogger = new PercentageLogger(logger,
+        (int) FileUtils.fetchNumberOfLines(result.rawResult()),
+        "Filling sheet" + sheet.getSheetName() + " with data...");
+    Files.readAllLines(result.rawResult())
         .forEach(
-            line -> fillRowWithData(sheet.createRow(rowIndex.getAndIncrement()), line, result));
+            line -> {
+              fillRowWithData(sheet.createRow(rowIndex.getAndIncrement()), line, result);
+              percentageLogger.incrementCounter();
+            });
   }
 
   private void fillRowWithData(Row row, String line, ScriptResult result) {
     AtomicInteger columnIndex = new AtomicInteger(0);
-    Arrays.stream(line.split(result.getCsvConfig().delimiter())).forEach(value -> {
-      row.createCell(columnIndex.getAndIncrement())
-          .setCellValue((value != null) ? value : ReporterConst.EMPTY_EXCEL_CELL);
-    });
+    Arrays.stream(line.split(result.csvConfig().delimiter())).forEach(
+        value -> row.createCell(columnIndex.getAndIncrement())
+            .setCellValue((value != null) ? value : ReporterConst.EMPTY_EXCEL_CELL));
   }
 
   private void autoSizeSheet(Sheet sheet) {
@@ -232,37 +257,53 @@ public class ReportGenerator {
     sheet.setAutoFilter(cra);
   }
 
-  private void addFormatToWorkbook(Workbook workbook, ReportTemplate template) {
+  private void addFormatToWorkbook(Workbook workbook, ReportTemplate template, Context context) {
     template.getSheetTemplates().forEach(sheetTemplate -> {
+      CellStyleContext cellStyleContext = new CellStyleContext(workbook);
       AtomicInteger columnNumber = new AtomicInteger(0);
+      logger.info("Adding general format to all cells...");
+      sheetTemplate.getFormatScripts().forEach(formatScript -> {
+        if (formatScript.getScript() != null) {
+          addFormatToAllCellsOfASheet(workbook, sheetTemplate, cellStyleContext, context,
+              formatScript.getScript());
+        }
+      });
+      logger.info("Adding column format to column cells...");
       sheetTemplate.getColumnTemplates().forEach(
           columnTemplate -> addFormatToWorkbook(workbook, sheetTemplate, columnTemplate,
-              columnNumber.getAndIncrement()));
+              columnNumber.getAndIncrement(), cellStyleContext, context));
     });
   }
 
   private void addFormatToWorkbook(Workbook workbook, SheetTemplate sheetTemplate,
-      ColumnTemplate columnTemplate, int columnNumber) {
+      ColumnTemplate columnTemplate, int columnNumber, CellStyleContext cellStyleContext,
+      Context context) {
     Sheet sheet = workbook.getSheet(sheetTemplate.getName());
-    addHeaderFormatToWorkbook(columnTemplate, sheet, columnNumber);
-    addValueFormatToWorkbook(columnTemplate, sheet, columnNumber);
+    logger.info("Adding header format to column '" + columnTemplate.getName() + "' in sheet '"
+        + sheetTemplate.getName() + "'...");
+    addHeaderFormatToWorkbook(columnTemplate, sheet, columnNumber, cellStyleContext, context);
+    logger.info("Adding value format to column '" + columnTemplate.getName() + "' in sheet '"
+        + sheetTemplate.getName() + "'...");
+    addValueFormatToWorkbook(columnTemplate, sheet, columnNumber, cellStyleContext, context);
   }
 
-  private void addHeaderFormatToWorkbook(ColumnTemplate template, Sheet sheet, int columnNumber) {
+  private void addHeaderFormatToWorkbook(ColumnTemplate template, Sheet sheet, int columnNumber,
+      CellStyleContext cellStyleContext, Context context) {
     if (template.getHeaderFormatScript() != null) {
       Script script = template.getHeaderFormatScript().getScript();
       if (script != null) {
-        fetchCellContext(script, sheet.getWorkbook()).applyCellStyleToCell(
+        fetchCellContext(script, cellStyleContext, context).applyCellStyleToCell(
             sheet.getRow(0).getCell(columnNumber));
       }
     }
   }
 
-  private void addValueFormatToWorkbook(ColumnTemplate template, Sheet sheet, int columnNumber) {
+  private void addValueFormatToWorkbook(ColumnTemplate template, Sheet sheet, int columnNumber,
+      CellStyleContext cellStyleContext, Context context) {
     if (template.getValueFormatScript() != null) {
       Script script = template.getValueFormatScript().getScript();
       if (script != null) {
-        CellContext cellContext = fetchCellContext(script, sheet.getWorkbook());
+        CellContext cellContext = fetchCellContext(script, cellStyleContext, context);
         int lastRowNum = sheet.getLastRowNum();
         if (lastRowNum > 0) {
           for (int i = 1; i <= lastRowNum; i++) {
@@ -273,9 +314,22 @@ public class ReportGenerator {
     }
   }
 
-  private CellContext fetchCellContext(Script script, Workbook workbook) {
+  private void addFormatToAllCellsOfASheet(Workbook workbook, SheetTemplate sheetTemplate,
+      CellStyleContext cellStyleContext, Context context, Script script) {
+    CellContext cellContext = fetchCellContext(script, cellStyleContext, context);
+    Sheet sheet = workbook.getSheet(sheetTemplate.getName());
+    PercentageLogger percentageLogger = new PercentageLogger(logger, sheet.getLastRowNum(),
+        "Adding format to all cells of sheet '" + sheetTemplate.getName() + "'...");
+    sheet.forEach(row -> {
+      row.forEach(cellContext::applyCellStyleToCell);
+      percentageLogger.incrementCounter();
+    });
+  }
+
+  private CellContext fetchCellContext(Script script, CellStyleContext cellStyleContext,
+      Context context) {
     try {
-      return scriptEngineManager.generateCellContext(script, workbook);
+      return scriptEngineManager.generateCellContext(script, cellStyleContext, context);
     } catch (ScriptEngineException e) {
       throw new RuntimeException(e);
     }