+++ src/main/java/de/samply/reporter/script/ScriptEngineManager.java
@@ -1,37 +1,49 @@
 package de.samply.reporter.script;
 
 import de.samply.reporter.context.CellContext;
+import de.samply.reporter.context.CellStyleContext;
 import de.samply.reporter.context.Context;
 import de.samply.reporter.template.ReportTemplate;
 import de.samply.reporter.template.script.Script;
 import de.samply.reporter.template.script.ScriptFramework;
+import de.samply.reporter.template.script.ScriptReference;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.poi.ss.usermodel.Workbook;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 @Component
 public class ScriptEngineManager {
 
-  private ScriptEngine[] scriptEngines = {new ThymeleafEngine(), new GroovyTemplatesEngine()};
-  private Map<ScriptFramework, ScriptEngine> idScriptEngineMap = new HashMap<>();
+  private final Logger logger = LoggerFactory.getLogger(ScriptEngineManager.class);
+  private final Map<ScriptFramework, ScriptEngine> idScriptEngineMap = new HashMap<>();
 
   public ScriptEngineManager() {
+    ScriptEngine[] scriptEngines = {new ThymeleafEngine(), new GroovyTemplatesEngine()};
     Arrays.stream(scriptEngines).forEach(
         scriptEngine -> idScriptEngineMap.put(scriptEngine.getScriptFramework(), scriptEngine));
   }
 
-  public Map<Script, ScriptResult> generateRawQualityReport(ReportTemplate template,
+  public Map<Script, ScriptResult> generateRawReport(ReportTemplate template,
       Context context) {
     Map<Script, ScriptResult> results = new HashMap<>();
+    addScriptToResultsAndGenerateRawResult(template.getInitScript(), results, context);
     template.getSheetTemplates().forEach(sheetTemplate -> {
-      if (sheetTemplate.getValuesScript() != null) {
-        Script script = sheetTemplate.getValuesScript().getScript();
-        results.put(script, generateRawResult(script, context));
-      }
+      logger.info(
+          "Generating temporal file with raw data for template " + sheetTemplate.getName() + "...");
+      addScriptToResultsAndGenerateRawResult(sheetTemplate.getValuesScript(), results, context);
     });
     return results;
+  }
+
+  private void addScriptToResultsAndGenerateRawResult(ScriptReference scriptReference,
+      Map<Script, ScriptResult> results, Context context) {
+    if (scriptReference != null) {
+      Script script = scriptReference.getScript();
+      results.put(script, generateRawResult(script, context));
+    }
   }
 
   private ScriptResult generateRawResult(Script script, Context context) {
@@ -58,8 +70,11 @@ public class ScriptEngineManager {
     return scriptFramework;
   }
 
-  public CellContext generateCellContext(Script script, Workbook workbook) throws ScriptEngineException {
-    return idScriptEngineMap.get(fetchScriptFramework(script)).generateCellContext(script, workbook);
+  public CellContext generateCellContext(Script script, CellStyleContext cellStyleContext,
+      Context context)
+      throws ScriptEngineException {
+    return idScriptEngineMap.get(fetchScriptFramework(script))
+        .generateCellContext(script, cellStyleContext, context);
   }
 
 }