+++ src/main/java/de/samply/reporter/script/ThymeleafEngine.java
@@ -2,6 +2,7 @@ package de.samply.reporter.script;
 
 import de.samply.reporter.app.ReporterConst;
 import de.samply.reporter.context.CellContext;
+import de.samply.reporter.context.CellStyleContext;
 import de.samply.reporter.context.Context;
 import de.samply.reporter.template.script.Script;
 import de.samply.reporter.template.script.ScriptFramework;
@@ -9,14 +10,13 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.StringWriter;
 import java.io.Writer;
-import org.apache.poi.ss.usermodel.Workbook;
 import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.templatemode.TemplateMode;
 import org.thymeleaf.templateresolver.StringTemplateResolver;
 
 public class ThymeleafEngine extends ScriptEngineImpl {
 
-  private TemplateEngine templateEngine = createTemplateEngine();
+  private final TemplateEngine templateEngine = createTemplateEngine();
 
   @Override
   public ScriptFramework getScriptFramework() {
@@ -34,18 +34,19 @@ public class ThymeleafEngine extends ScriptEngineImpl {
   }
 
   @Override
-  public CellContext generateCellContext(Script script, Workbook workbook)
-      throws ScriptEngineException {
-    CellContext cellContext = new CellContext(workbook);
+  public CellContext generateCellContext(Script script, CellStyleContext cellStyleContext,
+      Context context) {
+    CellContext cellContext = new CellContext(cellStyleContext);
     org.thymeleaf.context.Context thymeleafContext = new org.thymeleaf.context.Context();
-    thymeleafContext.setVariable(ReporterConst.CONTEXT_VARIABLE, cellContext);
+    thymeleafContext.setVariable(ReporterConst.CONTEXT_VARIABLE, context);
+    thymeleafContext.setVariable(ReporterConst.CELL_CONTEXT_VARIABLE, cellStyleContext);
     generateResult(script, thymeleafContext, new StringWriter());
     return cellContext;
   }
 
   private void generateResult(ScriptResult result, Script script,
       org.thymeleaf.context.Context context) throws ScriptEngineException {
-    try (FileWriter fileWriter = new FileWriter(result.getRawResult().toFile())) {
+    try (FileWriter fileWriter = new FileWriter(result.rawResult().toFile())) {
       generateResult(script, context, fileWriter);
     } catch (IOException e) {
       throw new ScriptEngineException(e);